Update README.md
This program implements a simple calendar event manager that uses HashMap to associate dates with a corresponding list of events. It can add and delete events, query the list of events on a specified date, check whether an event exists, and count the number of events on a given day.